Add testId props to components to explicitly call it out as an optional yet supported prop (#944)
* feat: add an explicit testId prop to components to encourage the use of the data-testid attribute

To encourage component consumers to use the data-testid attribute, a testId prop is being added to
components. Usage of this attribute will ease writing more robust tests.

SO-5324

* style: fix several linting errors

There were several linting errors present in the components that were updated. This commit addresses
most of those issues.

SO-5324

* style: fix linting issues present in components

Several existing components are failing the linting process. This addresses those issues.

SO-5324

* SO-5324: Bump the version.